added favorites and is_favorite fields to store details for displaying correct favorite and unfavorite buttons on frontend
